+++ perf/remote_thr.c
@@ -61,7 +61,7 @@ int main (int argc, char *argv [])
 
     for (i = 0; i != count; i++) {
         nbytes = nn_send (s, buf, sz, 0);
-        assert (nbytes == sz);
+        assert (nbytes == (int)sz);
     }
 
     free (buf);
